chore(stats): move stats part in another controller for visibility
